.walk() should respect "detail=True"
This PR addresses two issues: passing `detail` as a kwarg to
`.walk()` fails because it` internally calls `.ls()`
with both `detail=True` and whatever kwargs were passed
originally, causing `detail` to be specified twice.

The second is that `.walk()` does not respect `detail=True`
and always returns bare file names. My fix here is to just
shove the dict from `.ls()` into the return list unmodified --
this feels slightly incongruous with the non-detail version
which strips prefixes from paths, but this feels like the
right thing to do?

Anywhoo, this is a bit messy so I'm happy to clean it up
if you have any suggestions. It could also use some tests:
is there any particular way you'd like it tested? A test in
`test_local`?