REF: Refactor AbstractFileSysem init
This PR changes how init works for the base AbstractFileSystem. The
primary goal is to simplify things, which will hopefully result in more
predictable behavior downstream.

The primary goal is to ensure that the subclass's `__int__` is run when
the instance is newly created. The `__init__` should be skipped when
loading a cached instance. This is acomplished with a metaclass.

Additionally, I've split the alias handling into a secondary class. I
don't think dynamically adding methods in the `__init__` is a good, and
it created circular references which prevented garbage collection.

This should be backwards compatible. There is a large *behavior* change
with respect to caching though. On master, instances would never be
garbage collected unless the cache was explicitly cleared. We can easily
keep that behavior if we change `_cache` to a regular dict rather than a
WeakValueDictionary.