RF/BF: http - do not _fetch_all if file is smaller than blocksize/requested
This should prevent _fetch_all to overload the .cache with AllBytes
within async_fetch_all and thus use originally provided .cache and
store file on disk.

It is just a partial remedy/workaround to
https://github.com/intake/filesystem_spec/issues/553
and a proper solution is likely needing RFing so async_fetch_all could
feed the original .cache with the data for all blocks.