Add __eq__ and __hash__ for CachingFileSystem
Add equality testing and hashing for CachingFileSystem.  Add to_json()
raising an NotImplementedError and mark the corresponding test as xfail.
Implementing this is a bit more involved for me as I've never used
JSON before.  Although this might break code, it can only break code
that was already broken, as the current JSON representation is wrong.