avoid isfile calls on find as much as possible
If `walk` is also returning the same file, it should already have
been included in the `out`. If `out` is not empty, we can
safely assume that it is either a directory or a file.

If it's empty, it could be that it is either an empty directory
or, the path is a file and was not included from the `walk` results.
This time we will make a `isfile` call. This is unfortunate but
better than before, where we were making an additional for each find.

This was noticed in DVC, which was making remote cache querying slower
when traversing through the caches (512 calls instead of 256 :( ).