memfs: add global_store flag
By default, all instances of `MemoryFileSystem` use the same global store,
which has a few implications for users. For example:

* "memory leaks" if forgetting to delete files after use
* degrading performance (e.g. `info()` that looks through all entries in
  `store` and  `pseudo_dirs`) if storing 1K+ files (our use-case)

This PR allows one to choose whether to use a global store or not, allowing
for easy cleanup and more managable performance implications.

Having thousands of files stored in memfs would still result in a significant
performance hit, especially with complex tree structure. For cases like that,
we might want to create a trie-based memfs, that would be slightly slower to
build, but should be much faster and easier to traverse and read. We don't
have a need for this just yet, but we've ran into this problem in scmrepo's
gitfs, where we are using pygtrie ([1]).

[1] https://github.com/iterative/scmrepo/blob/main/scmrepo/fs.py