Catch PermissionError in fsspec.core when attempting to auto_mkdir
Catch `PermissionError` in fsspec.core when attempting to auto_mkdir for parent dir, as the user may not have permission to create the parent dir. 

This should hopefully allow the user to carry on without error if the parent dir already exists. If not, a later exception should be thrown when trying to write to a file when the parent dir doesn't exist.

This attempts to address the following issue:
* https://github.com/fsspec/filesystem_spec/issues/1404

Comments on how to improve this PR are welcome, as this is my first time contributing to the repo.