Fix tqdm callback garbage collection
Calling `close()` on `TqdmCallback.tqdm` reliably results in an exception
on interpreter exit. Debug prints show that in such cases, the value of
`tqdm` is not, as expected, the progress bar instance, but rather the
`no_op` bound method, which does not have a `close` attribute.

This change checks attribute existence of `close()` first before calling
it on the class. Checking this way does not require a `tqdm` import.