Fix non-str case in `AbstractFileSystem.expand_path()`
If the input to `expand_path` is not a string, it needs to be iterable,
since the if-branch for the container case tries to loop over the contents.

Previously, the code would fall through to the iterable case even for
perfectly legal inputs to `AbstractFileSystem._strip_protocol()`, like
`pathlib.Path`s, causing a crash with a less-than-ideal error message.

All of these inputs are now legalized by a changed `isinstance` check on
the scalar case, which now contains all types legal in `_strip_protocol`
and by proxy `stringify_path`.