Fix performance issues with MemoryFileSystem.rm
Closes #1724.

The idea is that the nominal case is deleting an existing file, and verifying that a
path corresponds to a file is an extremely fast key-existence check on the underlying
store (dict). Thus by checking first if it's a file we escape the expensive exists check,
which must scan all keys to check if the path is a prefix directory for some file.