Create an event loop if it doesn't exist
Python 3.14 has removed the implicit event loop creation. I was running
the test suite in Fedora test environment (copr) and in the buildsystem,
with varying results: sometimes the tests passed, more often they
didn't. I think that if the tests are run in parallel, sometimes they
reuse an event loop created for another test case. This ensures that if
no event loop is present, a new one will be created.