Prevent returning cached entry if the entry is degenerate
When `DatabricksFileSystem.ls` is called with a directory, it will query the API and
build a directory listing, then cache it.

If a subsequent `ls` is called on a directory that is child to the first directory,
the results that are returned are for the child directory itself, as though it were
just a file. The correct behavior is to get the contents of this child direcctory.

This PR ignores such responses so that `ls` will call the API again with the child
directory to get its contents. These results are then cached as usual and available
for future repeated `ls` calls with the child directory.