Add missing open-modes for memory filesystem (fixes #1921)
Memory filesystem's open() method supports some - but not all - file modes.
Specifically it supports "r+b" but not "a+b", "w+b", and "x+b". Add them.

(It also doesn't respect read-only/write-only properties; but the fix is more
involved and not (yet?) in demand; left as an exercise to the reader...)